Crate linera_wasm_test_runner

A binary for running unit tests for Linera applications implemented in WebAssembly.

The tests must be compiled for the wasm32-unknown-unknown target, and each unit test should use the webassembly_test attribute.

System APIs are not available to tests, so attempts to use any of them will lead to test failures.

Using With Cargo

It’s possible to make Cargo use this test runner for the wasm32-unknown-unknown target automatically by adding the following line to one of Cargo’s configuration files or setting the CARGO_TARGET_WASM32_UNKNOWN_UNKNOWN_RUNNER environment variable.

[target.wasm32-unknown-unknown]
runner = "/path/to/linera-wasm-test-runner"

Modules

mock_system_api 🔒

Structs

Options 🔒

Test 🔒

Information of a test in the input Wasm module.

TestReport 🔒

Summarizer of test results.

Functions

load_test_module 🔒

Loads the input test Wasm module specified as a command line argument.

main 🔒

Loads a test Wasm module and runs all test functions annotated by [webassembly-test].