Module linera_wasm_test_runner::mock_system_api

Structs

Resources

A map of resources allocated on the host side.

Functions

add_to_linker

Adds the mock system APIs to the linker, so that they are available to guest WebAsembly modules.

copy_memory_slices 🔒

Copies data from the source_offset to the destination_offset inside the guest WebAssembly module’s memory.

get_function 🔒

Retrieves a function exported from the guest WebAssembly module.

load_bytes 🔒

Loads a vector of length bytes starting at offset from the WebAssembly module’s memory.

load_indirect_bytes 🔒

Loads a vector of bytes with its starting offset and length stored in the WebAssembly module’s memory.

store_bytes_from_resource 🔒

Stores some bytes from a host-side resource to the WebAssembly module’s memory.

store_bytes_in_memory 🔒

Stores some bytes at the offset of the guest WebAssembly module’s memory.

store_in_memory 🔒

Stores a value at the offset of the guest WebAssembly module’s memory.