Struct linera_wasm_test_runner::Test

pub(crate) struct Test<'a> {
    pub(crate) name: &'a str,
    pub(crate) function: &'a str,
    pub(crate) ignore: bool,
}

Information of a test in the input Wasm module.

Fields
name: &'a strfunction: &'a strignore: bool
Implementations

impl<'a> Test<'a>

pub fn new(export: ExportType<'a>) -> Option<Self>

Collects test information from a function exported from the Wasm module.

pub async fn run( self, report: &mut TestReport, linker: &Linker<Resources>, test_module: &Module ) -> Result<()>

Runs a test function exported from the Wasm module and reports its result.

The test is executed in a clean Wasm environment.